+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [0ver](https://0ver.org) (more or less).
 
 ## [Unreleased]
 
+## [0.0.2] - 2020-06-29
+
 ### Added
 
 - New `ttl`, `min-ttl` and `force-generate` flags on the **get** function to manipulate credentials lengths
@@ -27,4 +29,6 @@ and this project adheres to [0ver](https://0ver.org) (more or less).
 - LICENSE
 - README
 
-[Unreleased]: https://github.com/mvisonneau/vac/compare/0.0.1...HEAD
+[Unreleased]: https://github.com/mvisonneau/vac/compare/0.0.2...HEAD
+[0.0.2]: https://github.com/mvisonneau/vac/tree/0.0.2
+[0.0.1]: https://github.com/mvisonneau/vac/tree/0.0.1